Created BasePVArray class for non OrderedPVArray specific methods
